requirements.txt
ZhipuAI
python-dotenv